Move operation binding to separate file
Operation binding in AdapterExtensionFeature has been moved to its own source file for better separation of concerns.
